rename oath-parser file and update README to be consistent
